refactor: 기존 constants.ts 경로를 리팩터링한 경로로 변경
