+++ src/features/announcement/components/ScheduleSettings/ScheduleSettings.tsx
@@ -4,7 +4,7 @@ import styles from "./ScheduleSettings.module.scss";
 import type { Schedule, ScheduleType } from "@/features/announcement/types/schedule";
 import { TimePicker } from "@/components/TimePicker/TimePicker";
 import { useAnnouncementStore } from "@/features/announcement/stores/useAnnouncementStore";
-import { ANNOUNCEMENT_DEFS } from "@/constants";
+import { ANNOUNCEMENT_DEFS } from "@/features/announcement/announcements";
 
 export function ScheduleSettings(): React.ReactNode {
   const { openSettingsId, schedules, updateSchedule, toggleSettings } = useAnnouncementStore();
